+++ examples/distributed_third_level/TaskAdvection.hpp
@@ -235,7 +235,7 @@ class TaskAdvection : public Task {
     dfg_->gatherFullGrid(fg, r);
   }
 
-  DistributedFullGrid<CombiDataType>& getDistributedFullGrid(int n = 0) override { return *dfg_; }
+  DistributedFullGrid<CombiDataType>& getDistributedFullGrid(size_t n = 0) override { return *dfg_; }
 
   void setZero() override {
     dfg_->setZero();
